+++ crates/xmtp_macro/src/async_trait.rs
@@ -1,0 +1,19 @@
+use quote::quote;
+
+/// A proc macro attribute that wraps the input in an `async_trait` implementation,
+/// delegating to the appropriate `async_trait` implementation based on the target architecture.
+///
+/// On wasm32 architecture, it delegates to `async_trait::async_trait(?Send)`.
+/// On all other architectures, it delegates to `async_trait::async_trait`.
+pub fn async_trait(
+    _attr: proc_macro::TokenStream,
+    input: proc_macro::TokenStream,
+) -> proc_macro::TokenStream {
+    let input = syn::parse_macro_input!(input as syn::Item);
+    quote! {
+        #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+        #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+        #input
+    }
+    .into()
+}
